requirements.txt
aiosmtplib==1.1.4
alembic==1.5.2
appnope==0.1.2
APScheduler==3.7.0
astroid==2.4.2
-e git+git@github.com:mr55p-dev/PageMail.git@8bf990a1b3354f88211010991633477d1de5a107#egg=async_scheduler&subdirectory=pythonAsyncScheduler
asyncpg==0.21.0
attrs==20.3.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cryptography==3.3.1
databases==0.4.1
decorator==4.4.2
dnspython==2.1.0
ecdsa==0.14.1
email-validator==1.1.2
fastapi==0.63.0
future==0.18.2
future-fstrings==1.2.0
git-pylint-commit-hook==2.5.1
gunicorn==20.0.4
h11==0.12.0
httptools==0.1.1
idna==2.10
iniconfig==1.1.1
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
isort==5.7.0
jedi==0.18.0
jupyter-client==6.1.12
jupyter-core==4.7.1
lazy-object-proxy==1.4.3
makefun==1.9.5
Mako==1.1.4
MarkupSafe==1.1.1
mccabe==0.6.1
networkx==2.5
packaging==20.9
parso==0.8.2
passlib==1.7.4
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.18
psycopg2==2.8.6
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pycparser==2.20
pydantic==1.7.3
Pygments==2.8.1
PyJWT==2.0.0
pylint==2.6.0
pyparsing==2.4.7
pytest==6.2.2
pytest-depends==1.0.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
python-jose==3.2.0
python-multipart==0.0.5
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.0.3
requests==2.25.1
rope==0.18.0
rsa==4.7
six==1.15.0
soupsieve==2.1
SQLAlchemy==1.3.22
starlette==0.13.6
toml==0.10.2
tornado==6.1
traitlets==5.0.5
tzlocal==2.1
urllib3==1.26.3
uvicorn==0.13.3
uvloop==0.14.0
watchgod==0.6
wcwidth==0.2.5
websockets==8.1
wrapt==1.12.1
